Add files via upload
I haven't added comments on everything, but i wanted to upload all files so you guys can try it on your end. Please note, this is a prototype and not all buttons have a bind. I also want to adjust the color and font.. along with the size proportions of certain things