Transaction page/Cart updated
The transaction page has a totals section on top (split into 2, so we can use one to calculate totals), takes user name and email, takes a cash input, provides change (the input field can be changed), has a section for credit card id input, has a click to pay button, and approval section (set as input but we can remove if needed). Last, we have a click to generate receipt at the bottom.